Extract scheduleEntry to a separate component
